add textHandler.js
